+++ humility/source/cli/changes.md
@@ -2,30 +2,78 @@ changes
 
 ---
 
-- read code inside the folder
-- commander is oop styled. i want a fp one. be declarative.
+done
 
-  ```
-  declaration: name desc commands options...
+commander is the defacto standard of creating cli.
 
-  hello_world = cli(declaration)
+i want to make it
 
-  parse(cli, args) # pretty print, make commands bold
+- fp
+- styled
 
-  help(cli, (command), terminal width){
-    string.
-  }
-  ```
+fp:
 
-- create a hello world cli, support these
-  - version (-v, --version), help (-h, --help)
-  - hello
-  - hello joe
-  - hello 'joe'
-  - hello --name 'joe'
+i dont like chaining.
 
-be as simple as possible. just make every command i want work as expected. no validation stuff.
+design an object to define a cli. write into spec. i will confirm.
+
+style:
+
+current
+
+```
+ ...humility/source/cli % cd "/home/fira/Documents/f/humility/source/cli/temp/" && br 'commander.ts'
+Usage: greeter [options] [command]
+
+An advanced greeting utility for production environments
+
+Options:
+  -V, --version            output the version number
+  -q, --quiet              Suppress all console logs except errors (default:
+                           false)
+  -h, --help               display help for command
+
+Commands:
+  person [options] <name>  Greet an individual person with customized styles
+  team [options]           Greet the entire development team
+  help [command]           display help for command
+```
+
+expected
+
+```
+An advanced greeting utility for production environments
+
+**Usage**: **greeter** [command] [options]
+
+**Commands**:
+  **person** [options] <name>  Greet an individual person with customized styles
+  **team** [options]           Greet the entire development team
+  **help** [command]           display help for command
+
+**Options**:
+  **-V**, **--version**            output the version number
+  **-q**, **--quiet**              Suppress all console logs except errors (default:
+                           false)
+  **-h**, **--help**               display help for command
+```
+
+use chalk to make titles, commands, and flags bold.
+
+usage:
+
+```
+import {parse} from cli
+
+greet = {
+  
+}
+
+parse(greet, process args) # output
+```
+
+build on top of commander
+
+write cli.ts and greet.ts
 
 ---
-
-